feat: create write review rating bar
