+++ app/src/main/java/com/caffeine/capin/review/PictureUriEntity.kt
@@ -1,0 +1,4 @@
+package com.caffeine.capin.review
+
+class PictureUriEntity {
+}
