+++ app/src/main/java/com/caffeine/capin/review/WriteReviewActivity.kt
@@ -1,0 +1,4 @@
+package com.caffeine.capin.review
+
+class WriteReviewActivity {
+}
